// src/generated/types.gen.ts
export type Object = {
    type: string;
    [key: string]: unknown;
};
// ...
export type AccountLinks = {
    avatar?: Link;
    [key: string]: unknown;
};

/**
 * Link
 *
 * A link to a resource related to this object.
 */
export type Link = {
    href?: string;
    name?: string;
};

export type Account = Omit<Object, 'type'> & {
    links?: AccountLinks;
    created_on?: string;
    display_name?: string;
    uuid?: string;
    type: 'account';
    [key: string]: unknown;
};
// ...
export type Comment = Omit<Object, 'type'> & {
    id?: number;
    created_on?: string;
    updated_on?: string;
    content?: {
        /**
         * The text as it was typed by a user.
         */
        raw?: string;
        /**
         * The type of markup language the raw content is to be interpreted in.
         */
        markup?: 'markdown' | 'creole' | 'plaintext';
        /**
         * The user's content rendered as HTML.
         */
        html?: string;
    };
    user?: Account;
    deleted?: boolean;
    parent?: Comment;
    inline?: {
        /**
         * The comment's anchor line in the old version of the file. If the comment is a multi-line comment, this is the ending line number in the old version of the file.
         */
        from?: number;
        /**
         * The comment's anchor line in the new version of the file. If the comment is a multi-line comment, this is the ending line number in the new version of the file.
         */
        to?: number;
        /**
         * The starting line number in the old version of the file, if the comment is a multi-line comment. This is null otherwise.
         */
        start_from?: number;
        /**
         * The starting line number in the new version of the file, if the comment is a multi-line comment. This is null otherwise.
         */
        start_to?: number;
        /**
         * The path of the file this comment is anchored to.
         */
        path: string;
    };
    links?: {
        /**
         * Link
         *
         * A link to a resource related to this object.
         */
        self?: {
            href?: string;
            name?: string;
        };
        /**
         * Link
         *
         * A link to a resource related to this object.
         */
        html?: {
            href?: string;
            name?: string;
        };
        /**
         * Link
         *
         * A link to a resource related to this object.
         */
        code?: {
            href?: string;
            name?: string;
        };
    };
    type: 'comment';
    [key: string]: unknown;
};
// ...
export type CommentResolution = {
    type: string;
    user?: Account;
    /**
     * The ISO8601 timestamp the resolution was created.
     */
    created_on?: string;
    [key: string]: unknown;
};
// ...
export type PullrequestComment = Omit<Comment, 'type'> & {
    pullrequest?: Pullrequest;
    resolution?: CommentResolution;
    pending?: boolean;
    type: 'pullrequest_comment';
    [key: string]: unknown;
};